+++ src/CWITC.DataObjects/Feedback.cs
@@ -7,6 +7,6 @@
     {
         public string UserId { get; set; }
         public string SessionId { get; set; }
-        public int SessionRating { get; set; }
+        public double SessionRating { get; set; }
     }
 }
